Server/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
pymysql
prometheus-client
